# openapi.yaml
openapi: 3.0.1
info:
  title: Wayfarer
  description: "An  online booking platform that provides transportation to individuals. This helps solve problems of quick transportation."
  version: "1.0.0"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "orlandonkwoji@gmail.com"
  license:
    name: "MIT"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: http://localhost:3700/api/v1
    description: Development server
...
paths:
...
  /trips:
    post:
      tags:
      - Trip
      summary: Create a trip.
      description: >
        A `trip` created for clients booked for a journey from `origin` to `destination`
      operationId: createTrip
      requestBody:
        $ref: '#/components/requestBodies/createTripBody'
      responses:
        201:
          $ref: '#/components/responses/createTripResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    required:
                      - message
                    properties:
                      message:
                        type: string
                        example: 'destination: is a required field'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          description: Conflict
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    required:
                      - message
                    properties:
                      message:
                        type: string
                        example: This bus is unavailable
        409:
          description: Conflict
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    required:
                      - message
                    properties:
                      message:
                        type: string
                        example: An active trip with this bus exists
        500:
          $ref: '#/components/responses/Unexpected'                
    get:
      tags:
      - Trip
      summary: Get trips.
      description: >
         An array of all `trips` is returned to the client
      operationId: getTrips
      parameters:
        - name: origin
          in: query
          schema:
            type: string
          description: >
            Specifies the starting location of a `trip`.
            This parameter is incompatible with `destination`.
        - name: destination
          in: query
          schema:
            type: string
          description: >
            Specifies the end location of a `trip`.
            This parameter is incompatible with `origin`.
      responses:
        200:
          $ref: '#/components/responses/getTripsResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    required:
                      - message
                    properties:
                      message:
                        type: string
                        example: '"origin" is incompatible with "destination"'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/Unexpected'                
...
  /bookings:
    post:
      tags:
      - Booking
      summary: Creates a booking.
      description: >
        A booking for a `trip` is created
      operationId: createBooking
      requestBody:
        $ref: '#/components/requestBodies/CreateBookingBody'
      responses:
        201:
          $ref: '#/components/responses/SuccessCreateBookingResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    required:
                      - message
                    properties:
                      message:
                        type: string
                        example: 'trip_id: is a required field'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          description: Conflict
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    required:
                      - message
                    properties:
                      message:
                        type: string
                        example: This trip is unavailable
        409:
          description: Conflict
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    required:
                      - message
                    properties:
                      message:
                        type: string
                        example: You are already booked on this trip
        500:
          $ref: '#/components/responses/Unexpected'
    get:
      tags:
      - Booking
      summary: Get bookings.
      description: >
         An array of all `bookings` is returned
      operationId: getBookings
      responses:
        200:
          $ref: '#/components/responses/GetBookingsResponse'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/Unexpected'                
  /bookings/{booking_id}:
    delete:
      tags:
      - Booking
      summary: Delete a booking.
      description: >
        A `booking` is deleted by the client
      operationId: deleteBooking
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            exclusiveMinimum: false
          description: The booking ID
      responses:
        200:
          $ref: '#/components/responses/DeleteBookingResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    required:
                      - message
                    properties:
                      message:
                        type: string
                        example: 'booking_id: is a required field'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          description: Conflict
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    required:
                      - message
                    properties:
                      message:
                        type: string
                        example: Booking not found
        500:
          $ref: '#/components/responses/Unexpected'                
components:
  responses:
...
    SuccessCreateBookingResponse:
      description: A JSON response that contains the booking object
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessObject'
              - type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Booking'
      links:
        TakeBusIdToCreateTrip:
          $ref: '#/components/links/TakeBusIdToCreateTrip'
    createTripResponse:
      description: A JSON response that contains the trip object
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessObject'
              - type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Trip'
...
    DeleteBookingResponse:
      description: A JSON response that contains the booking object
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessObject'
              - type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/DeleteBooking'
    getTripsResponse:
      description: A JSON response that contains an array of trips object
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessArray'
              - type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/TripsArray'
    GetBookingsResponse:
      description: A JSON response that contains an array of bookings object
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessArray'
              - type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/BookingsArray'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    enum: 
                      - Invalid token or none provided
    Forbidden:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    enum: 
                      - Access Denied
    Unexpected:
      description: Unexpected
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                required:
                  - message
                properties:
                  message:
                    type: string
  requestBodies:
...
    CreateBookingBody:
      description: A JSON object containing the booking credentials
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - trip_id
            properties:
              trip_id:
                allOf:
                  - $ref: '#/components/schemas/IdInt'
                  - type: integer
                    example: 1
              seat_number:
                allOf:
                  - $ref: '#/components/schemas/IdInt'
                  - type: integer
                    example: 2
    createTripBody:
      description: A JSON object containing the trip credentials
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Trip'
  schemas:
...
    Error:
      type: object
      properties:
        status:
          type: string
          enum: ['error']
        message:
          type: string
      required:
        - status
        - message
    # Schema for success response body where data is an object
    SuccessObject:
      type: object
      properties:
        status:
          type: string
          enum: ['success']
        data:
          type: object
      required:
        - status
        - data
    # Schema for success response body where data is an array
    SuccessArray:
      type: object
      properties:
        status:
          type: string
          enum: ['success']
        data:
          type: array
          items:
            type: object
      required:
        - status
        - data
...
    Booking: 
      title: Booking
      type: object
      required:
        - booking_id
        - user_id
        - trip_id
        - bus_id
        - trip_date 
        - seat_number
        - first_name
        - last_name
        - email
      properties:
        booking_id:
          allOf:
            - $ref: '#/components/schemas/IdInt'
            - type: integer
              readOnly: true
              example: 1
        user_id:
          allOf:
            - $ref: '#/components/schemas/IdInt'
            - type: integer
              readOnly: true
              example: 2
        trip_id:
          allOf:
            - $ref: '#/components/schemas/IdInt'
            - type: integer
              readOnly: true
              example: 1
        bus_id:
          allOf:
            - $ref: '#/components/schemas/IdInt'
            - type: integer
              readOnly: true
              example: 1
        trip_date:
          allOf:
            - type: string
              format: date
              example: '2020-05-12'
            - type: string
              pattern: 'YYYY/MM/DD'
        seat_number:
          allOf:
            - $ref: '#/components/schemas/IdInt'
            - type: integer
              example: 2
        first_name:
          type: string
          example: ORLANDO
        last_name:
            type: string
            example: NKWOJI
        email:
          type: string
          format: email
          example: orlando@wayfarer.com
    Trip: 
      title: Trip
      type: object
      required:
        - trip_id
        - bus_id
        - origin
        - destination
        - trip_date
        - fare
        - status
      properties:
        trip_id:
          allOf:
            - $ref: '#/components/schemas/IdInt'
            - type: integer
              readOnly: true
              example: 2
        bus_id:
          allOf:
            - $ref: '#/components/schemas/IdInt'
            - type: integer
              example: 1
        origin:
          allOf:
            - $ref: '#/components/schemas/Alphanumeric'
            - type: string
              example: LAGOS
        destination:
          allOf:
            - $ref: '#/components/schemas/Alphanumeric'
            - type: string
              example: ABUJA
        trip_date:
          allOf:
            - type: string
              format: date
              example: '2020-05-12'
            - type: string
              pattern: 'YYYY/MM/DD'
        fare:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
          example: 500
        status:
          type: string
          enum: [active, cancelled]
          default: active
...
    DeleteBooking: 
      allOf:
        - type: object
          required:
            - message
          properties:
            message:
              type: string
              enum: ['Booking deleted successfully']
        - $ref: '#/components/schemas/Booking'
    TripsArray: 
      title: TripsArray
      type: array
      items:
        $ref: '#/components/schemas/Trip'
      minItems: 0
      uniqueItems: true
    BookingsArray: 
      title: BookingsArray
      type: array
      items:
        $ref: '#/components/schemas/Booking'
      minItems: 0
      uniqueItems: true
    IdInt: 
      title: ID
      type: integer
      format: int32
      minimum: 1
      exclusiveMinimum: false
    Alphanumeric: 
      title: Alphanumeric
      type: string
      pattern: '^[A-Z0-9]{3,16}$'
...
  links:
    TakeBusIdToCreateTrip:
      description: >
        The `bus_id` value returned in the response can be used as the `bus_id` in the `requestBody` of `POST /trips`.
      operationId: createTrip
      requestBody:
        bus_id: '$response.body#/data/bus_id'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []